dalgo_chat_dashboard/requirements.txt
streamlit==1.39.0
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.17
psycopg2-binary==2.9.9
chromadb==0.5.23
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
pyyaml==6.0.2